Update README.md
What is FAKE USDT SENDER SOFTWARE?* 

This is a software that can generate and send flash usdt to any wallet of your choice and the fusdt perform and works just like the real usdt for 21 days and after the 21 days the fake usdt dissappear from the wallet and the software generate up to 500k daily. It's a lifetime software which works on phone and PC, and remember it supports all wallet. 

 

*Features Of fake usdt sender software:* 

*Fake Usdt sender software Core network version 10.0.1*

 

```Features Of The Software``` 

-Transaction Last 21days. 

 

-100% User Anonymity

-Can cancel  anytime

-Support all wallet

- USABILITY : 1USER

-TRANSFERABLE

-PROXY : TOR/SOCKS/MPLS.

_ Daily limit: $500k

 https://officialfakeusdtsender.com/